LPS-179177 Extract CommerceCountryRegionsScreenNavigationCategory methods to a CommerceCountryRegionsScreenNavigationEntry component and extends CommerceCountryRegionsScreenNavigationCategory in it
