+++ modules/apps/commerce/commerce-address-web/src/main/java/com/liferay/commerce/address/web/internal/frontend/taglib/servlet/taglib/CommerceCountryRegionsScreenNavigationCategory.java
@@ -15,32 +15,12 @@
 package com.liferay.commerce.address.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.commerce.address.web.internal.constants.CommerceCountryScreenNavigationConstants;
-import com.liferay.commerce.address.web.internal.display.context.CommerceRegionsDisplayContext;
-import com.liferay.commerce.address.web.internal.portlet.action.helper.ActionHelper;
-import com.liferay.commerce.constants.CommerceConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.Country;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.resource.PortletResourcePermission;
-import com.liferay.portal.kernel.service.RegionService;
-import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -49,23 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=20"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceCountryRegionsScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<Country> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return CommerceCountryScreenNavigationConstants.
-			CATEGORY_KEY_COMMERCE_COUNTRY_REGIONS;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return CommerceCountryScreenNavigationConstants.
 			CATEGORY_KEY_COMMERCE_COUNTRY_REGIONS;
 	}
@@ -75,7 +46,7 @@ public class CommerceCountryRegionsScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, "regions");
+		return language.get(resourceBundle, "regions");
 	}
 
 	@Override
@@ -84,61 +55,7 @@ public class CommerceCountryRegionsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_COUNTRY_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, Country country) {
-		if (country == null) {
-			return false;
-		}
-
-		return true;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		RenderRequest renderRequest =
-			(RenderRequest)httpServletRequest.getAttribute(
-				JavaConstants.JAVAX_PORTLET_REQUEST);
-		RenderResponse renderResponse =
-			(RenderResponse)httpServletRequest.getAttribute(
-				JavaConstants.JAVAX_PORTLET_RESPONSE);
-
-		CommerceRegionsDisplayContext commerceRegionsDisplayContext =
-			new CommerceRegionsDisplayContext(
-				_actionHelper, _portletResourcePermission, _regionService,
-				renderRequest, renderResponse);
-
-		renderRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT, commerceRegionsDisplayContext);
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/commerce_country/commerce_regions.jsp");
-	}
-
 	@Reference
-	private ActionHelper _actionHelper;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(resource.name=" + CommerceConstants.RESOURCE_NAME_COMMERCE_ADDRESS + ")"
-	)
-	private PortletResourcePermission _portletResourcePermission;
-
-	@Reference
-	private RegionService _regionService;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.address.web)"
-	)
-	private ServletContext _servletContext;
+	protected Language language;
 
 }
